# openapi.yml
openapi: 3.0.3
info:
  title: Space Marine Management API
  version: 1.0.0
  description: API for managing Space Marines, their Coordinates, and Chapters with full CRUD, pagination support, special operations, embedded entities, and export capabilities.
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
...
  /coordinates:
    get:
      summary: Get all Coordinates (paginated)
      operationId: getAllCoordinates
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
      responses:
        "200":
          description: Paginated list of Coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCoordinatesResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create new Coordinates
      operationId: createCoordinates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Coordinates"
      responses:
        "201":
          description: Coordinates created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoordinatesWithId"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /coordinates/{id}:
    get:
      summary: Get Coordinates by ID
      operationId: getCoordinatesById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Coordinates found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoordinatesWithId"
        "404":
          description: Coordinates not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update Coordinates by ID
      operationId: updateCoordinates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Coordinates"
      responses:
        "200":
          description: Coordinates updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoordinatesWithId"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Coordinates not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete Coordinates by ID
      operationId: deleteCoordinates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Coordinates deleted
        "404":
          description: Coordinates not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /coordinates/export/json:
    get:
      summary: Export Coordinates as JSON
      operationId: exportCoordinatesJson
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 1000
          description: Number of items per page (max 1000 for exports)
      responses:
        "200":
          description: JSON file download with paginated Coordinates data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoordinatesWithId"
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "coordinates_page_0_size_1000.json"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /coordinates/export/xml:
    get:
      summary: Export Coordinates as XML
      operationId: exportCoordinatesXml
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 1000
          description: Number of items per page (max 1000 for exports)
      responses:
        "200":
          description: XML file download with paginated Coordinates data
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoordinatesWithId"
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "coordinates_page_0_size_1000.xml"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # =========
  # CHAPTERS
  # =========
...
components:
  schemas:
...
    PaginatedCoordinatesResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CoordinatesWithId"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
...
    Coordinates:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
          format: int64
          description: X coordinate value
        y:
          type: number
          format: float
          maximum: 343
          description: Y coordinate value (must be ≤ 343)
    CoordinatesWithId:
      allOf:
        - $ref: "#/components/schemas/Coordinates"
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              format: int64
              minimum: 1
              description: Unique auto-generated ID

    # === CHAPTER ===
...
    ErrorResponse:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
      example:
        error: "Invalid weapon types: LASER_CANNON. Valid values: BOLTGUN,HEAVY_BOLTGUN,FLAMER,HEAVY_FLAMER,MULTI_MELTA"
        timestamp: "2025-11-07T14:30:00.000Z"